(CU-85zt0qwa2 ui) migrate: ui utils to ui modules
